+++ pom.xml
@@ -131,6 +131,12 @@
 		    <artifactId>knative-model</artifactId>
 		    <version>5.5.0</version>
 		</dependency>
+	    <dependency>
+	      <groupId>io.fabric8</groupId>
+	      <artifactId>kubernetes-server-mock</artifactId>
+		    <version>5.5.0</version>
+	      <scope>test</scope>
+	    </dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-devtools</artifactId>
