+++ src/main/java/io/boomerang/kube/service/KubeServiceImpl.java
@@ -8,6 +8,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.ConfigMapBuilder;
@@ -22,6 +23,7 @@ import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientException;
 
 @Component
+@Primary
 public class KubeServiceImpl implements KubeService {
 
   private static final Logger LOGGER = LogManager.getLogger(KubeServiceImpl.class);
@@ -73,7 +75,7 @@ public class KubeServiceImpl implements KubeService {
   @Value("${controller.service.host}")
   protected String controllerServiceURL;
 
-  KubernetesClient client = null;
+  protected KubernetesClient client = null;
 
   protected KubeServiceImpl() {
     this.client = new DefaultKubernetesClient();
