+++ src/main/resources/application-local.properties
@@ -1,5 +1,8 @@
 logging.level.root=INFO
 
+# Fabric8 client configuration
+kubernetes.namespace=default
+
 # Disable elastic integration locally
 kube.worker.logging.type=default
 kube.worker.node.dedicated=true
