+++ src/test/java/io/boomerang/tests/service/CICDControllerServiceTest.java
@@ -1,17 +1,16 @@
-package io.boomerang.service;
+package io.boomerang.tests.service;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
-
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
-
 import javax.servlet.http.HttpServletResponse;
 
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mockito;
@@ -23,12 +22,12 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;
 import io.boomerang.kube.exception.KubeRuntimeException;
-import io.boomerang.kube.service.FlowKubeServiceImpl;
 import io.boomerang.model.Response;
 import io.boomerang.model.Storage;
 import io.boomerang.model.Task;
 import io.boomerang.model.TaskResponse;
 import io.boomerang.model.Workflow;
+import io.boomerang.tests.kube.CICDKubeServiceImpl;
 import io.kubernetes.client.ApiException;
 import io.kubernetes.client.models.V1ConfigMap;
 import io.kubernetes.client.models.V1Job;
@@ -38,173 +37,216 @@ import io.kubernetes.client.models.V1Status;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest
-@ActiveProfiles("live")
-public class FlowControllerServiceTest {
+@ActiveProfiles("cicd")
+public class CICDControllerServiceTest {
 
   @Autowired
-  private FlowControllerServiceImpl flowControllerService;
+  private CICDControllerServiceImpl cicdControllerService;
 
   @MockBean
-  private FlowKubeServiceImpl kubeService;
+  private CICDKubeServiceImpl kubeService;
 
   @Test
   public void testCreateWorkflow() throws ApiException {
     Workflow workflow = getDefaultWorkflow();
     Storage storage = new Storage();
     storage.setEnable(true);
-    storage.setSize("200");
     workflow.setWorkflowStorage(storage);
 
     V1PersistentVolumeClaim claim = new V1PersistentVolumeClaim();
     V1PersistentVolumeClaimStatus status = new V1PersistentVolumeClaimStatus();
     status.setPhase("phase1");
-    Mockito.when(kubeService.createPVC("workflowName", "workflowId", "workflowActivityId", "200"))
+    Mockito.when(kubeService.createPVC("workflowName", "workflowId", "workflowActivityId", null))
         .thenReturn(claim);
     Mockito.when(kubeService.watchPVC("workflowId", "workflowActivityId")).thenReturn(status);
 
-    Mockito.when(kubeService.createWorkflowConfigMap("workflowName", "workflowId",
-        "workflowActivityId", null)).thenReturn(new V1ConfigMap());
-    Mockito.when(kubeService.watchConfigMap("workflowId", "workflowActivityId", null))
-        .thenReturn(new V1ConfigMap());
-
-    Response response = flowControllerService.createWorkflow(workflow);
+    Response response = cicdControllerService.createWorkflow(workflow);
     assertEquals("0", response.getCode());
-    assertTrue(response.getMessage().startsWith("Workflow Activity (workflowActivityId"));
+    assertTrue(response.getMessage().startsWith("Component Activity (workflowActivityId"));
 
-    Mockito.verify(kubeService).createPVC("workflowName", "workflowId", "workflowActivityId",
-        "200");
+    Mockito.verify(kubeService).createPVC("workflowName", "workflowId", "workflowActivityId", null);
     Mockito.verify(kubeService).watchPVC("workflowId", "workflowActivityId");
-    Mockito.verify(kubeService).createWorkflowConfigMap("workflowName", "workflowId",
-        "workflowActivityId", null);
-    Mockito.verify(kubeService).watchConfigMap("workflowId", "workflowActivityId", null);
+  }
+
+  @Test
+  public void testCreateWorkflowWithStorageDisabled() throws ApiException {
+    Workflow workflow = getDefaultWorkflow();
+    Storage storage = new Storage();
+    storage.setEnable(false);
+    workflow.setWorkflowStorage(storage);
+
+    Response response = cicdControllerService.createWorkflow(workflow);
+    assertEquals("0", response.getCode());
+    assertTrue(response.getMessage().startsWith("Component Activity (workflowActivityId"));
+
+    Mockito.verifyNoMoreInteractions(kubeService);
   }
 
   @Test
   public void testCreateWorkflowWithException() throws ApiException {
     Workflow workflow = getDefaultWorkflow();
     Storage storage = new Storage();
-    storage.setEnable(false);
-    storage.setSize("200");
+    storage.setEnable(true);
     workflow.setWorkflowStorage(storage);
-    workflow.setProperties(new HashMap<>());
-    
-    Mockito.when(kubeService.createWorkflowConfigMap("workflowName", "workflowId",
-        "workflowActivityId", workflow.getProperties())).thenThrow(KubeRuntimeException.class);
 
-    Response response = flowControllerService.createWorkflow(workflow);
+    Mockito.when(kubeService.createPVC("workflowName", "workflowId", "workflowActivityId", null))
+        .thenThrow(KubeRuntimeException.class);
+
+    Response response = cicdControllerService.createWorkflow(workflow);
     assertEquals("1", response.getCode());
     assertTrue(response.getMessage()
         .startsWith("net.boomerangplatform.kube.exception.KubeRuntimeException"));
 
-    Mockito.verify(kubeService).createWorkflowConfigMap("workflowName", "workflowId",
-        "workflowActivityId", workflow.getProperties());
+    Mockito.verify(kubeService).createPVC("workflowName", "workflowId", "workflowActivityId", null);
   }
 
   @Test
   public void testTerminateWorkflow() {
     Workflow workflow = getDefaultWorkflow();
-    Mockito.when(kubeService.deletePVC(workflow.getWorkflowId(), workflow.getWorkflowActivityId()))
-        .thenReturn(new V1Status());
-    Mockito.when(kubeService.deleteConfigMap(workflow.getWorkflowId(),
-        workflow.getWorkflowActivityId(), null)).thenReturn(new V1Status());
+    Mockito.when(kubeService.deletePVC(workflow.getWorkflowId(), null)).thenReturn(new V1Status());
 
-    Response response = flowControllerService.terminateWorkflow(workflow);
+    Response response = cicdControllerService.terminateWorkflow(workflow);
     assertEquals("0", response.getCode());
-    assertTrue(response.getMessage().startsWith("Workflow Activity (workflowActivityId"));
+    assertTrue(response.getMessage().startsWith("Component Storage (workflowId"));
 
-    Mockito.verify(kubeService).deletePVC(workflow.getWorkflowId(),
-        workflow.getWorkflowActivityId());
-    Mockito.verify(kubeService).deleteConfigMap(workflow.getWorkflowId(),
-        workflow.getWorkflowActivityId(), null);
+    Mockito.verify(kubeService).deletePVC(workflow.getWorkflowId(), null);
   }
 
   @Test
   public void testTerminateWorkflowWithException() throws ApiException {
     Workflow workflow = getDefaultWorkflow();
-    Mockito.when(kubeService.deletePVC(workflow.getWorkflowId(), workflow.getWorkflowActivityId()))
+    Mockito.when(kubeService.deletePVC(workflow.getWorkflowId(), null))
         .thenThrow(KubeRuntimeException.class);
 
-    Response response = flowControllerService.terminateWorkflow(workflow);
+    Response response = cicdControllerService.terminateWorkflow(workflow);
     assertEquals("1", response.getCode());
     assertTrue(response.getMessage()
         .startsWith("net.boomerangplatform.kube.exception.KubeRuntimeException"));
 
-    Mockito.verify(kubeService).deletePVC(workflow.getWorkflowId(),
-        workflow.getWorkflowActivityId());
+    Mockito.verify(kubeService).deletePVC(workflow.getWorkflowId(), null);
   }
 
   @Test
   public void testExecuteTask() {
     Task task = getDefaultTask();
 
+    Mockito.when(kubeService.checkPVCExists(task.getWorkflowId(), null, null, true))
+        .thenReturn(false);
     Mockito.when(kubeService.createTaskConfigMap(task.getWorkflowName(), task.getWorkflowId(),
         task.getWorkflowActivityId(), task.getTaskName(), task.getTaskId(),
         task.getProperties())).thenReturn(new V1ConfigMap());
-    Mockito.when(kubeService.watchConfigMap(null, task.getWorkflowActivityId(), task.getTaskId()))
-        .thenReturn(new V1ConfigMap());
+    Mockito.when(kubeService.watchConfigMap(task.getWorkflowId(), task.getWorkflowActivityId(),
+        task.getTaskId())).thenReturn(new V1ConfigMap());
     Mockito.when(kubeService.createJob(task.getWorkflowName(), task.getWorkflowId(),
-        task.getWorkflowActivityId(),task.getTaskActivityId(), task.getTaskName(), task.getTaskId(), task.getArguments(),
+        task.getWorkflowActivityId(), task.getTaskActivityId(), task.getTaskName(), task.getTaskId(), task.getArguments(),
         task.getProperties())).thenReturn(new V1Job());
     Mockito.when(
         kubeService.watchJob(task.getWorkflowId(), task.getWorkflowActivityId(), task.getTaskId()))
         .thenReturn(new V1Job());
 
-    Mockito
-        .when(kubeService.getTaskOutPutConfigMapData(task.getWorkflowId(),
-            task.getWorkflowActivityId(), task.getTaskId(), task.getTaskName()))
-        .thenReturn(new HashMap<>());
-    Mockito.when(kubeService.deleteConfigMap(null, task.getWorkflowActivityId(), task.getTaskId()))
-        .thenReturn(new V1Status());
+    Mockito.when(kubeService.deleteConfigMap(task.getWorkflowId(), task.getWorkflowActivityId(),
+        task.getTaskId())).thenReturn(new V1Status());
 
-    TaskResponse response = flowControllerService.executeTask(task);
+    TaskResponse response = cicdControllerService.executeTask(task);
     assertEquals("0", response.getCode());
     assertTrue(response.getMessage().startsWith("Task (taskId) has been executed successfully."));
-    assertNotNull(response.getOutput());
 
+    Mockito.verify(kubeService).checkPVCExists(task.getWorkflowId(), null, null, true);
     Mockito.verify(kubeService).createTaskConfigMap(task.getWorkflowName(), task.getWorkflowId(),
         task.getWorkflowActivityId(), task.getTaskName(), task.getTaskId(),
         task.getProperties());
-    Mockito.verify(kubeService).watchConfigMap(null, task.getWorkflowActivityId(),
+    Mockito.verify(kubeService).watchConfigMap(task.getWorkflowId(), task.getWorkflowActivityId(),
         task.getTaskId());
     Mockito.verify(kubeService).createJob(task.getWorkflowName(), task.getWorkflowId(),
-        task.getWorkflowActivityId(), task.getTaskActivityId(),task.getTaskName(), task.getTaskId(), task.getArguments(),
+    	task.getWorkflowActivityId(), task.getTaskActivityId(), task.getTaskName(), task.getTaskId(), task.getArguments(),
         task.getProperties());
     Mockito.verify(kubeService).watchJob(task.getWorkflowId(), task.getWorkflowActivityId(),
         task.getTaskId());
 
-    Mockito.verify(kubeService).getTaskOutPutConfigMapData(task.getWorkflowId(),
-        task.getWorkflowActivityId(), task.getTaskId(), task.getTaskName());
-    Mockito.verify(kubeService).deleteConfigMap(null, task.getWorkflowActivityId(),
+    Mockito.verify(kubeService).deleteConfigMap(task.getWorkflowId(), task.getWorkflowActivityId(),
         task.getTaskId());
   }
 
   @Test
-  public void testExecuteTaskWithException() {
+  @Ignore 
+  public void testExecuteTaskWithCache() throws ApiException {
     Task task = getDefaultTask();
-    task.setArguments(new ArrayList<>());
+    task.setProperty("component/cache.enabled", "true");
 
+    Mockito.when(kubeService.checkPVCExists(task.getWorkflowId(), null, null, true))
+        .thenReturn(false);
+    Mockito.when(kubeService.createPVC(task.getWorkflowName(), task.getWorkflowId(),
+        task.getWorkflowActivityId(), null)).thenReturn(new V1PersistentVolumeClaim());
+    Mockito.when(kubeService.watchPVC(task.getWorkflowId(), task.getWorkflowActivityId()))
+        .thenReturn(new V1PersistentVolumeClaimStatus());
     Mockito.when(kubeService.createTaskConfigMap(task.getWorkflowName(), task.getWorkflowId(),
         task.getWorkflowActivityId(), task.getTaskName(), task.getTaskId(),
-        task.getProperties())).thenThrow(KubeRuntimeException.class);
-    Mockito
-        .when(kubeService.getTaskOutPutConfigMapData(task.getWorkflowId(),
-            task.getWorkflowActivityId(), task.getTaskId(), task.getTaskName()))
-        .thenReturn(new HashMap<>());
-    Mockito.when(kubeService.deleteConfigMap(null, task.getWorkflowActivityId(), task.getTaskId()))
-        .thenReturn(new V1Status());
+        task.getProperties())).thenReturn(new V1ConfigMap());
+    Mockito.when(kubeService.watchConfigMap(task.getWorkflowId(), task.getWorkflowActivityId(),
+        task.getTaskId())).thenReturn(new V1ConfigMap());
+    Mockito.when(kubeService.createJob(task.getWorkflowName(), task.getWorkflowId(),
+        task.getWorkflowActivityId(), task.getTaskActivityId(), task.getTaskName(), task.getTaskId(), task.getArguments(),
+        
+        task.getProperties())).thenReturn(new V1Job());
+    Mockito.when(
+        kubeService.watchJob(task.getWorkflowId(), task.getWorkflowActivityId(), task.getTaskId()))
+        .thenReturn(new V1Job());
 
-    TaskResponse response = flowControllerService.executeTask(task);
-    assertEquals("1", response.getCode());
-    assertTrue(response.getMessage()
-        .startsWith("net.boomerangplatform.kube.exception.KubeRuntimeException"));
-    assertNotNull(response.getOutput());
+    Mockito.when(kubeService.deleteConfigMap(task.getWorkflowId(), task.getWorkflowActivityId(),
+        task.getTaskId())).thenReturn(new V1Status());
+
+    TaskResponse response = cicdControllerService.executeTask(task);
+    assertEquals("0", response.getCode());
+    assertTrue(response.getMessage().startsWith("Task (taskId) has been executed successfully."));
+
+    Mockito.verify(kubeService).checkPVCExists(task.getWorkflowId(), null, null, true);
+    Mockito.verify(kubeService).createPVC(task.getWorkflowName(), task.getWorkflowId(),
+        task.getWorkflowActivityId(), null);
+    Mockito.verify(kubeService).watchPVC(task.getWorkflowId(), task.getWorkflowActivityId());
 
     Mockito.verify(kubeService).createTaskConfigMap(task.getWorkflowName(), task.getWorkflowId(),
         task.getWorkflowActivityId(), task.getTaskName(), task.getTaskId(),
         task.getProperties());
-    Mockito.verify(kubeService).getTaskOutPutConfigMapData(task.getWorkflowId(),
-        task.getWorkflowActivityId(), task.getTaskId(), task.getTaskName());
-    Mockito.verify(kubeService).deleteConfigMap(null, task.getWorkflowActivityId(),
+    Mockito.verify(kubeService).watchConfigMap(task.getWorkflowId(), task.getWorkflowActivityId(),
+        task.getTaskId());
+    Mockito.verify(kubeService).createJob(task.getWorkflowName(), task.getWorkflowId(),
+        task.getWorkflowActivityId(), task.getTaskActivityId(), task.getTaskName(), task.getTaskId(), task.getArguments(),
+        task.getProperties());
+    Mockito.verify(kubeService).watchJob(task.getWorkflowId(), task.getWorkflowActivityId(),
+        task.getTaskId());
+
+    Mockito.verify(kubeService).deleteConfigMap(task.getWorkflowId(), task.getWorkflowActivityId(),
+        task.getTaskId());
+  }
+
+  
+  
+  @Test
+  public void testExecuteTaskWithException() {
+    Task task = getDefaultTask();
+    task.setArguments(new ArrayList<>());
+    task.setArgument("argument");
+
+    Mockito.when(kubeService.checkPVCExists(task.getWorkflowId(), null, null, true))
+        .thenReturn(true);
+    Mockito.when(kubeService.deletePVC(task.getWorkflowId(), null)).thenReturn(new V1Status());
+    Mockito.when(kubeService.createTaskConfigMap(task.getWorkflowName(), task.getWorkflowId(),
+        task.getWorkflowActivityId(), task.getTaskName(), task.getTaskId(),
+        task.getProperties())).thenThrow(KubeRuntimeException.class);
+    Mockito.when(kubeService.deleteConfigMap(task.getWorkflowId(), task.getWorkflowActivityId(),
+        task.getTaskId())).thenReturn(new V1Status());
+
+    TaskResponse response = cicdControllerService.executeTask(task);
+    assertEquals("1", response.getCode());
+    assertTrue(response.getMessage()
+        .startsWith("net.boomerangplatform.kube.exception.KubeRuntimeException"));
+
+    Mockito.verify(kubeService).checkPVCExists(task.getWorkflowId(), null, null, true);
+    Mockito.verify(kubeService).deletePVC(task.getWorkflowId(), null);
+    Mockito.verify(kubeService).createTaskConfigMap(task.getWorkflowName(), task.getWorkflowId(),
+        task.getWorkflowActivityId(), task.getTaskName(), task.getTaskId(),
+        task.getProperties());
+    Mockito.verify(kubeService).deleteConfigMap(task.getWorkflowId(), task.getWorkflowActivityId(),
         task.getTaskId());
   }
 
@@ -219,7 +261,7 @@ public class FlowControllerServiceTest {
     Mockito.doNothing().when(kubeService).patchTaskConfigMap(workflowId, workflowActivityId, taskId,
         taskName, properties);
 
-    Response response = flowControllerService.setTaskResultParameter(workflowId, workflowActivityId,
+    Response response = cicdControllerService.setTaskResultParameter(workflowId, workflowActivityId,
         taskId, taskName, "key", "value");
     assertEquals("0", response.getCode());
     assertTrue(response.getMessage().startsWith(
@@ -240,7 +282,7 @@ public class FlowControllerServiceTest {
     Mockito.doThrow(KubeRuntimeException.class).when(kubeService).patchTaskConfigMap(workflowId,
         workflowActivityId, taskId, taskName, properties);
 
-    Response response = flowControllerService.setTaskResultParameter(workflowId, workflowActivityId,
+    Response response = cicdControllerService.setTaskResultParameter(workflowId, workflowActivityId,
         taskId, taskName, "key", "value");
     assertEquals("1", response.getCode());
     assertTrue(response.getMessage()
@@ -261,7 +303,7 @@ public class FlowControllerServiceTest {
     Mockito.doNothing().when(kubeService).patchTaskConfigMap(workflowId, workflowActivityId, taskId,
         taskName, properties);
 
-    Response response = flowControllerService.setTaskResultParameters(workflowId, workflowActivityId,
+    Response response = cicdControllerService.setTaskResultParameters(workflowId, workflowActivityId,
         taskId, taskName, properties);
     assertEquals("0", response.getCode());
     assertTrue(response.getMessage().startsWith(
@@ -282,7 +324,7 @@ public class FlowControllerServiceTest {
     Mockito.doThrow(KubeRuntimeException.class).when(kubeService).patchTaskConfigMap(workflowId,
         workflowActivityId, taskId, taskName, properties);
 
-    Response response = flowControllerService.setTaskResultParameters(workflowId, workflowActivityId,
+    Response response = cicdControllerService.setTaskResultParameters(workflowId, workflowActivityId,
         taskId, taskName, properties);
     assertEquals("1", response.getCode());
     assertTrue(response.getMessage()
@@ -301,7 +343,7 @@ public class FlowControllerServiceTest {
     Mockito.when(kubeService.getPodLog(workflowId, workflowActivityId, taskId,taskId))
         .thenReturn("Success");
 
-    Response response = flowControllerService.getLogForTask(workflowId, workflowActivityId, taskId,taskId);
+    Response response = cicdControllerService.getLogForTask(workflowId, workflowActivityId, taskId,taskId);
     assertEquals("0", response.getCode());
     assertEquals("Success", response.getMessage());
 
@@ -317,7 +359,7 @@ public class FlowControllerServiceTest {
     Mockito.when(kubeService.getPodLog(workflowId, workflowActivityId, taskId,taskId))
         .thenThrow(KubeRuntimeException.class);
 
-    Response response = flowControllerService.getLogForTask(workflowId, workflowActivityId, taskId,taskId);
+    Response response = cicdControllerService.getLogForTask(workflowId, workflowActivityId, taskId,taskId);
     assertEquals("1", response.getCode());
     assertTrue(response.getMessage()
         .startsWith("net.boomerangplatform.kube.exception.KubeRuntimeException"));
@@ -336,7 +378,7 @@ public class FlowControllerServiceTest {
         .thenReturn(mock(StreamingResponseBody.class));
 
     StreamingResponseBody streamingResponseBody =
-        flowControllerService.streamLogForTask(response, workflowId, workflowActivityId, taskId,taskId);
+        cicdControllerService.streamLogForTask(response, workflowId, workflowActivityId, taskId,taskId);
     assertNotNull(streamingResponseBody);
 
     Mockito.verify(kubeService).streamPodLog(response, workflowId, workflowActivityId, taskId,taskId);
@@ -353,10 +395,18 @@ public class FlowControllerServiceTest {
         .thenThrow(KubeRuntimeException.class);
 
     StreamingResponseBody streamingResponseBody =
-        flowControllerService.streamLogForTask(response, workflowId, workflowActivityId, taskId,taskId);
+        cicdControllerService.streamLogForTask(response, workflowId, workflowActivityId, taskId,taskId);
     assertNull(streamingResponseBody);
 
     Mockito.verify(kubeService).streamPodLog(response, workflowId, workflowActivityId, taskId,taskId);
+  }
+
+  private Workflow getDefaultWorkflow() {
+    Workflow workflow = new Workflow();
+    workflow.setWorkflowActivityId("workflowActivityId");
+    workflow.setWorkflowName("workflowName");
+    workflow.setWorkflowId("workflowId");
+    return workflow;
   }
 
   private Task getDefaultTask() {
@@ -370,13 +420,4 @@ public class FlowControllerServiceTest {
     task.setProperty("name1", "value1");
     return task;
   }
-
-  private Workflow getDefaultWorkflow() {
-    Workflow workflow = new Workflow();
-    workflow.setWorkflowActivityId("workflowActivityId");
-    workflow.setWorkflowName("workflowName");
-    workflow.setWorkflowId("workflowId");
-    return workflow;
-  }
-
 }
