--- src/test/resources/application-cicd.properties
@@ -1,15 +1,0 @@
-spring.profiles.active=local
-
-kube.api.type=custom
-kube.api.base.path=https://9.42.30.59:8001
-kube.api.token=eyJhbGciOiJSUzI1NiIsImtpZCI6IiJ9.eyJpc3MiOiJrdWJlcm5ldGVzL3NlcnZpY2VhY2NvdW50Iiwia3ViZXJuZXRlcy5pby9zZXJ2aWNlYWNjb3VudC9uYW1lc3BhY2UiOiJrdWJlLXN5c3RlbSIsImt1YmVybmV0ZXMuaW8vc2VydmljZWFjY291bnQvc2VjcmV0Lm5hbWUiOiJibXJnLWNpY2QtdXJiYW5jb2RlLXRva2VuLXdia3pwIiwia3ViZXJuZXRlcy5pby9zZXJ2aWNlYWNjb3VudC9zZXJ2aWNlLWFjY291bnQubmFtZSI6ImJtcmctY2ljZC11cmJhbmNvZGUiLCJrdWJlcm5ldGVzLmlvL3NlcnZpY2VhY2NvdW50L3NlcnZpY2UtYWNjb3VudC51aWQiOiIyZjRkMTQ3My0zZjVlLTExZTktOTI0Ni0wMDUwNTZhMGY0MGMiLCJzdWIiOiJzeXN0ZW06c2VydmljZWFjY291bnQ6a3ViZS1zeXN0ZW06Ym1yZy1jaWNkLXVyYmFuY29kZSJ9.jVeR3nLG_XSsWYHVjxjAmS_-fLEWby6MvclHVwgtx3ixMFDidA5ugn5i7KSaBtwT1_VZu_VShuLFgZ1t97Uxb7r2OU5K-lHJFKUYzGVwCYrGolsEMa4ul2-l7DK9SiWdINVDq9ggvC0hPAq4GbyuP9cZ0FMQqMzqyIuP1p2J0hPOsnH0gZs1cFxfgM2peuUzOGUlDqTaVfHf9zCpbM5nlRvUJYcXQajj7Ug3IBne8LxQ2aKL-meLr4_3_jmpgn6Ov-lIb8iUfeKmHnQBW14BQKvKLEbkGNRld_4TOE2oE_UQCnrvX4-lE0OMfgiWJprbZs3hvtvmgd7ar2PAUEq3nw
-kube.api.debug=true
-kube.namespace=default
-
-proxy.enable=false
-
-#swagger
-swagger.enable=true
-springfox.documentation.swagger.v2.path=/controller/api-docs
-
-opentracing.jaeger.enabled=false
